Revert "Move solver interface out of `__init__.py` (#103)"
This reverts commit 46b45ab58ade2c6750fa92526510d3b6982b958f.
