+++ tests/test_all.py
@@ -11,7 +11,6 @@ import gurobipy as gp
 import pytest
 
 import cvxpy_gurobi
-from cvxpy_gurobi.translation import CVXPY_VERSION
 from test_problems import ProblemTestCase
 from test_problems import all_valid_problems
 
@@ -62,7 +61,7 @@ def test_lp(case: ProblemTestCase, snapshot: SnapshotFixture, tmp_path: Path) ->
         )
     )
 
-    if CVXPY_VERSION[:2] == (1, 6):
+    if cvxpy_gurobi.CVXPY_VERSION[:2] == (1, 6):
         assert snapshot() == output
     else:
         # don't update snapshots nor delete them
