+++ tests/test_non_optimal.py
@@ -3,7 +3,6 @@ import gurobipy as gp
 from cvxpy import settings as s
 
 import cvxpy_gurobi
-from cvxpy_gurobi.translation import CVXPY_VERSION
 
 
 def test_backfill_unbounded() -> None:
@@ -11,7 +10,7 @@ def test_backfill_unbounded() -> None:
     cvxpy_gurobi.solve(problem)
     model = problem.solver_stats.extra_stats
     assert problem.status == s.UNBOUNDED
-    if CVXPY_VERSION >= (1, 2, 0):
+    if cvxpy_gurobi.CVXPY_VERSION >= (1, 2, 0):
         assert model.Status == gp.GRB.Status.UNBOUNDED
 
 
@@ -21,5 +20,5 @@ def test_backfill_infeasible() -> None:
     cvxpy_gurobi.solve(problem, **{gp.GRB.Param.DualReductions: 0})
     model = problem.solver_stats.extra_stats
     assert problem.status == s.INFEASIBLE
-    if CVXPY_VERSION >= (1, 2, 0):
+    if cvxpy_gurobi.CVXPY_VERSION >= (1, 2, 0):
         assert model.Status == gp.GRB.Status.INFEASIBLE
